requirements.txt
absl-py==0.8.1
docopt==0.6.2
h5py==2.9.0
matlabengineforpython===R2019b
matplotlib==3.1.1
multiprocess==0.70.9
numpy==1.17.2
pandas==0.25.3
phate==1.0.2
pickleshare==0.7.5
pyparsing==2.4.2
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.3.1
scprep==1.0.3
tqdm==4.30.0